replace mysql2 monkey patching with prepend
